fix(evalhub/kueue): resolve pre-commit issues and fixture conflicts
Fixes all remaining CodeRabbit review comments and pre-commit failures:

**Fixture Organization:**
- Moved evalhub_mt_* fixtures to parent evalhub/conftest.py
- Removed duplicate fixtures from kueue/ and multitenancy/ conftest.py
- Both subdirectories now inherit shared fixtures from parent
- Resolved pytest fixture name conflict (UFN001) errors

**Code Quality Fixes:**
- Fixed CSV readiness check: use csv.instance.status.phase instead of csv.status
- Narrowed try/except scope in kueue_unmanaged_dsc fixture
- Catches only during DSC component lookup, not during yield/teardown
- Fixed broad Exception catching: (AttributeError, KeyError) specific exceptions
- Renamed evalhub_kueue_rbac_ready → evalhub_kueue_tenant_rbac (noun not adjective)

**Test Cleanup:**
- Removed model_namespace parametrization from TestEvalHubKueueBasic
- Kueue-specific model_namespace fixture handles namespace creation with labels
- All 4 kueue tests collect successfully
- All 127 multitenancy tests collect successfully

**Pre-commit Status:**
- ✅ flake8: Passed (removed unused imports)
- ✅ ruff: Passed (auto-fixed 6 errors)
- ✅ ruff format: Passed (reformatted files)
- ✅ mypy: Passed
- ⚠️ check-prohibited-patterns: Pre-existing issue in backup_utils.py (not my changes)

All tests collect without fixture conflicts. Ready for CI.
